pkg/docker/config: Bring auth tests from docker/docker_client_test
Shifted a number of auth-getting unit tests from docker_client_test.go
into config_test.go, since they only excercise config.go logic.
They'd been in their previous location since landing in d30079fd (Be
benevolent to .docker/config.json file, 2016-09-20, #96).
